move check_output to utils
