+++ gensim/utils.py
@@ -34,6 +34,7 @@ import multiprocessing
 import shutil
 import sys
 from contextlib import contextmanager
+import subprocess
 
 import numpy
 import scipy.sparse
@@ -43,8 +44,6 @@ if sys.version_info[0] >= 3:
 
 from six import iteritems, u, string_types, unichr
 from six.moves import xrange
-
-from models.wrappers.wrapper_utils import check_output # wrapper_utils for external process management
 
 try:
     from smart_open import smart_open
@@ -1127,3 +1126,26 @@ def keep_vocab_item(word, count, min_count, trim_rule=None):
             return False
         else:
             return default_res
+
+def check_output(*popenargs, **kwargs):
+    r"""Run command with arguments and return its output as a byte string.
+    Backported from Python 2.7 as it's implemented as pure python on stdlib.
+    >>> check_output(['/usr/bin/python', '--version'])
+    Python 2.6.2
+    Added extra KeyboardInterrupt handling
+    """
+    try:
+        process = subprocess.Popen(stdout=subprocess.PIPE, *popenargs, **kwargs)
+        output, unused_err = process.communicate()
+        retcode = process.poll()
+        if retcode:
+            cmd = kwargs.get("args")
+            if cmd is None:
+                cmd = popenargs[0]
+            error = subprocess.CalledProcessError(retcode, cmd)
+            error.output = output
+            raise error
+        return output
+    except KeyboardInterrupt:
+        process.terminate()
+        raise
